requirements.txt
Flask
opencv-python-headless
numpy
tensorflow
Pillow
scikit-learn
